requirements.txt
apispec==3.3.2
attrs==22.1.0
Babel==2.11.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.4
Flask==2.2.2
Flask-AppBuilder==4.1.6
Flask-Babel==2.0.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.38.0
fuzzywuzzy==0.18.0
gunicorn==20.1.0
idna==3.4
inflect==6.0.2
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.6.2
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pluggy==1.0.0
prison==0.2.1
psycopg2-binary==2.9.5
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
PyTDC==0.3.8
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
rdkit-pypi==2022.9.1
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1